+++ packages/wallets/walletconnect/CHANGELOG.md
@@ -31,7 +31,7 @@
 
 ### Patch Changes
 
--   f5abf15c: Using WalletConnect (alone, or through `@solana/wallet-adapter-wallets` no longer fatals in Next 13.
+-   f5abf15c: Using WalletConnect (alone, or through `@solana/wallet-adapter-wallets-dtc` no longer fatals in Next 13.
 
 ## 0.1.11